Added new dto mapper for stock information

+++ src/main/kotlin/stock/me/provider/response/StockInformationDto.kt
@@ -5,8 +5,8 @@ data class StockInformationDto(
     val priceToEarnings : Double?,
     val earningsPerShare : Double?,
     val priceToBook: Double?,
-    val marketCap: Long?,
+    val marketCap: Double?,
     val sharesOwned: Long?,
-    val revenue: Long?,
+    val revenue: Double?,
     val shortRatio: Double?
 )
